refactor: Move some components in TabsHeader
Allows to override completely TabsHeader
